+++ pkg/apis/vsphere/v1/main_test.go
@@ -1,0 +1,13 @@
+package v1
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestVSphere(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Generic vSphere API tests")
+}